Added a lot, check description
Chat is now working with firebase and the JSQ messages framework. Has
bugs in that it will display ‘optional’ in the text.

Changed the look of course a little bit.

In account settings, you now have to meet some certain requirements in
the fields to be able to actually save. BUG -> When trying to access
cell a cell that is off screen, the app will crash

In new product it will actually save now! This updates to server and
everything. BUG -> The server will crash for 10 minutes when you upload
a product. Insane!